move fetchFaveStar methods into useFaveStar hook
