+++ superset-frontend/src/views/CRUD/dashboard/DashboardList.tsx
@@ -20,11 +20,7 @@ import { SupersetClient } from '@superset-ui/connection';
 import { t } from '@superset-ui/translation';
 import React, { useState, useMemo } from 'react';
 import rison from 'rison';
-import {
-  createFetchRelated,
-  createErrorHandler,
-  createFaveStarHandlers,
-} from 'src/views/CRUD/utils';
+import { createFetchRelated, createErrorHandler } from 'src/views/CRUD/utils';
 import { useListViewResource, useFavoriteStatus } from 'src/views/CRUD/hooks';
 import ConfirmStatusChange from 'src/components/ConfirmStatusChange';
 import SubMenu from 'src/components/Menu/SubMenu';
@@ -79,7 +75,13 @@ function DashboardList(props: DashboardListProps) {
     t('dashboard'),
     props.addDangerToast,
   );
-  const [favoriteStatusRef, setFavoriteStatus] = useFavoriteStatus({});
+  const [favoriteStatusRef, fetchFaveStar, saveFaveStar] = useFavoriteStatus(
+    {},
+    FAVESTAR_BASE_URL,
+    (message: string) => {
+      props.addDangerToast(message);
+    },
+  );
 
   const [dashboardToEdit, setDashboardToEdit] = useState<Dashboard | null>(
     null,
@@ -90,15 +92,6 @@ function DashboardList(props: DashboardListProps) {
   const canExport = hasPerm('can_mulexport');
 
   const initialSort = [{ id: 'changed_on_delta_humanized', desc: true }];
-
-  const fetchFaveStarMethods = createFaveStarHandlers(
-    FAVESTAR_BASE_URL,
-    favoriteStatusRef.current,
-    setFavoriteStatus,
-    (message: string) => {
-      props.addDangerToast(message);
-    },
-  );
 
   function openDashboardEditModal(dashboard: Dashboard) {
     setDashboardToEdit(dashboard);
@@ -174,8 +167,8 @@ function DashboardList(props: DashboardListProps) {
     return (
       <FaveStar
         itemId={id}
-        fetchFaveStar={fetchFaveStarMethods.fetchFaveStar}
-        saveFaveStar={fetchFaveStarMethods.saveFaveStar}
+        fetchFaveStar={fetchFaveStar}
+        saveFaveStar={saveFaveStar}
         isStarred={!!favoriteStatusRef.current[id]}
         height={20}
         width={20}
@@ -322,7 +315,7 @@ function DashboardList(props: DashboardListProps) {
         disableSortBy: true,
       },
     ],
-    [canEdit, canDelete, canExport],
+    [canEdit, canDelete, canExport, favoriteStatusRef],
   );
 
   const filters: Filters = [
