+++ superset-frontend/src/views/CRUD/utils.tsx
@@ -62,41 +62,3 @@ export function createErrorHandler(handleErrorFunc: (errMsg?: string) => void) {
     handleErrorFunc(parsedError.message);
   };
 }
-
-export function createFaveStarHandlers(
-  baseURL: string,
-  favoriteStatus: FavoriteStatus,
-  setFavoriteStatus: (update: Partial<FavoriteStatus>) => void,
-  handleErrorFunc: (message: string) => void,
-) {
-  const fetchFaveStar = (id: number) => {
-    SupersetClient.get({
-      endpoint: `${baseURL}/${id}/count/`,
-    })
-      .then(({ json }) => {
-        setFavoriteStatus({ [id]: json.count > 0 });
-      })
-      .catch(() =>
-        handleErrorFunc(t('There was an error fetching the favorite status')),
-      );
-  };
-
-  const saveFaveStar = (id: number, isStarred: boolean) => {
-    const urlSuffix = isStarred ? 'unselect' : 'select';
-
-    SupersetClient.get({
-      endpoint: `${baseURL}/${id}/${urlSuffix}/`,
-    })
-      .then(() => {
-        setFavoriteStatus({ [id]: !isStarred });
-      })
-      .catch(() =>
-        handleErrorFunc(t('There was an error saving the favorite status')),
-      );
-  };
-
-  return {
-    fetchFaveStar,
-    saveFaveStar,
-  };
-}
